+++ elasticsearch-20170613/core/Models/TurnOnZoneRequest.cs
@@ -9,6 +9,10 @@ using Tea;
 namespace AlibabaCloud.SDK.Elasticsearch20170613.Models
 {
     public class TurnOnZoneRequest : TeaModel {
+        [NameInMap("hpAlbZoneDrained")]
+        [Validation(Required=false)]
+        public bool? HpAlbZoneDrained { get; set; }
+
         /// <summary>
         /// <b>Example:</b>
         /// <para>cn-hangzhou-i</para>